rename macro.py file
